benchmarking/requirements.txt
torch
transformers
pydantic
requests
lm-format-enforcer
pandas
pyarrow
accelerate
scipy
python-dotenv
setuptools
